refactor(customers): change dto's folder
