+++ services/inventory/src/main/java/com/petek/inventoryService/dto/customer/CustomerResponse.java
@@ -1,4 +1,4 @@
-package com.petek.inventoryService.dto;
+package com.petek.inventoryService.dto.customer;
 
 import java.time.Instant;
