+++ services/inventory/src/main/java/com/petek/inventoryService/dto/customer/CustomerUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.petek.inventoryService.dto;
+package com.petek.inventoryService.dto.customer;
 
 import com.petek.inventoryService.entity.Customer.CustomerSegment;
